refactor: move season matches slice into core application
